[Rustify Network] Test works
