+++ rust/Cargo.toml
@@ -15,6 +15,8 @@ pyo3 = {version = "0.18.3", features = ["extension-module"]}
 zmq = "0.10.0"
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"
+rand = "0.5.0"
+bincode = "1.3.3"
 
 [target.x86_64-apple-darwin]
 rustflags = [
