+++ rust/src/network/mod.rs
@@ -1,3 +1,6 @@
 pub mod channel;
 pub mod io_loop;
 pub mod socket_meta;
+pub mod data_writer;
+pub mod data_reader;
+pub mod utils;
