+++ rust/src/network/utils.rs
@@ -1,0 +1,9 @@
+use rand::{distributions::Alphanumeric, Rng};
+
+pub fn random_string(len: usize) -> String {
+    rand::thread_rng()
+        .sample_iter(&Alphanumeric)
+        .take(len)
+        .map(char::from)
+        .collect()
+}
